fix: CustomerName에 최대 길이 검증 추가 & dto reqeust 패키지 컨트롤러로 이동
